src/task1/requirements.txt
pathlib
pandas
python-dotenv
requests
nptyping
uvicorn
fastapi
loguru
scikit-learn
